% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiprocessing.R, R/workspace.R
\name{.jmp.sa.count}
\alias{.jmp.sa.count}
\alias{count}
\alias{.jws.multiprocessing.count}
\title{Count the number of objects inside a workspace or multiprocessing}
\usage{
.jmp.sa.count(jmp)

.jws.multiprocessing.count(jws)
}
\arguments{
\item{jws, jmp}{the workspace or the multiprocessing.}
}
\description{
Functions to count the number of multiprocessing inside a workspace (`jws.multiprocessing.count`) or
the number of SaItem inside a multiprocessing (`jmp.sa.count`).
}
