% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiprocessing.R
\name{.jmp_name}
\alias{.jmp_name}
\title{Get the name of a multiprocessing or a SaItem}
\usage{
.jmp_name(jmp)
}
\arguments{
\item{jmp, jsa}{the object to retrieve the name from.}
}
\description{
Functions to retrieve the name of a multiprocessing (`jmp_name()`) or SaItem (`jsa_name()`).
}
