% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workspace.R
\name{.jws.multiprocessing}
\alias{.jws.multiprocessing}
\title{Extract a Multiprocessing or a SaItem}
\usage{
.jws.multiprocessing(jws, idx)
}
\arguments{
\item{jws, jmp}{the workspace or the multiprocessing.}

\item{idx}{index of the object to extract.}
}
\description{
Extract a Multiprocessing or a SaItem
}
