% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workspace.R
\name{.jws.open}
\alias{.jws.open}
\title{Load a 'JDemetra+' workpace}
\usage{
.jws.open(file)
}
\arguments{
\item{file}{the path to the 'JDemetra+' workspace to load.
By default a dialog box opens.}

\item{jws}{the workspace.}
}
\description{
`.jws.open()` loads a workspace and `.jws.compute()` computes it (to be able to get all the models).
}
\seealso{
[.jws.load()] to directly load a workspace and import all the models.
}
