% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tramoseats_spec.R
\name{jd2r_spec_tramo}
\alias{jd2r_spec_tramo}
\title{Title}
\usage{
jd2r_spec_tramo(spec, context_dictionary = NULL, extra_info = FALSE, freq = NA)
}
\arguments{
\item{freq}{}
}
\value{

}
\description{
Title
}
