% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/context.R
\name{jd3_utilities}
\alias{jd3_utilities}
\alias{jd2p_context}
\title{Java Utility Functions}
\usage{
jd2p_context(jcntx)
}
\arguments{
\item{jcntx}{parameters.}
}
\description{
These functions are used in all JDemetra+ 3.0 packages to easily interact between R and Java objects.
}
