% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saitem.R, R/multiprocessing.R
\name{jsa.name}
\alias{jsa.name}
\alias{jmp.name}
\title{Get the name of a multiprocessing or a SaItem}
\usage{
jsa.name(jsa)

jmp.name(jmp)
}
\arguments{
\item{jmp, jsa}{the object to retrieve the name from.}
}
\description{
Functions to retrieve the name of a multiprocessing (`jmp.name()`) or SaItem (`jsa.name()`).
}
