% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiprocessing.R, R/workspace.R
\name{jws.load}
\alias{jws.load}
\alias{jmp.load}
\title{Read all SaItems}
\usage{
jmp.load(jmp)

jws.load(file)
}
\arguments{
\item{jmp}{a multiprocessing.}

\item{file}{the path to the 'JDemetra+' workspace to load.
By default a dialog box opens.}
}
\description{
Functions to read all the SAItem (\code{\link[=jsa.read]{jsa.read()}}) of a multiprocessing (\code{jmp.load()})
or a workspace (\code{jws.load}).
}
