% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiprocessing.R, R/workspace.R
\name{jws.multiprocessing}
\alias{jws.multiprocessing}
\alias{jmp.sa}
\title{Extract a Multiprocessing or a SaItem}
\usage{
jmp.sa(jmp, idx)

jws.multiprocessing(jws, idx)
}
\arguments{
\item{idx}{index of the object to extract.}

\item{jws, jmp}{the workspace or the multiprocessing.}
}
\description{
Extract a Multiprocessing or a SaItem
}
