% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workspace.R
\name{load_workspace}
\alias{load_workspace}
\title{Read all SaItems}
\usage{
load_workspace(file)
}
\arguments{
\item{jmp}{a multiprocessing.}
}
\description{
Functions to read all the SAItem ([jsa.read()]) of a multiprocessing (`jmp.load()`)
or a workspace (`.jws.load()`).
}
