% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/x13.R
\name{regarima.refresh}
\alias{regarima.refresh}
\title{Title}
\usage{
regarima.refresh(
  spec,
  refspec = NULL,
  policy = c("FreeParameters", "Complete", "Outliers_StochasticComponent", "Outliers",
    "FixedParameters", "FixedAutoRegressiveParameters", "Fixed"),
  period = 0,
  start = NULL,
  end = NULL
)
}
\arguments{
\item{end}{}
}
\value{

}
\description{
Title
}
