% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spec_rjd.R
\name{spec_tramo_jd2r}
\alias{spec_tramo_jd2r}
\title{Title}
\usage{
spec_tramo_jd2r(
  spec = NA,
  context_dictionary = NULL,
  extra_info = FALSE,
  freq = NA
)
}
\arguments{
\item{freq}{}
}
\value{

}
\description{
Title
}
