% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tramoseats.R
\name{tramoseats_raw}
\alias{tramoseats_raw}
\title{Title}
\usage{
tramoseats_raw(ts, spec, context = NULL)
}
\arguments{
\item{ts}{Time series}

\item{spec}{Specification list}

\item{context}{Optional. context for the processing (user-defined variables and calendars)}
}
\value{

}
\description{
Title
}
