% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/x13.R
\name{x13}
\alias{x13}
\title{Title}
\usage{
x13(ts, spec, context = NULL)
}
\arguments{
\item{ts}{Time series}

\item{spec}{Specification list}

\item{context}{Optional. context for the processing (user-defined variables and calendars)}
}
\value{

}
\description{
Title
}
