% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workspace.R
\name{load_workspace}
\alias{load_workspace}
\title{Read all SaItems}
\usage{
load_workspace(file)
}
\arguments{
\item{file}{the path to the 'JDemetra+' workspace to load.
By default a dialog box opens.}

\item{jmp}{a multiprocessing.}
}
\description{
Functions to read all the SAItem ([jsa_read()]) of a multiprocessing (`jmp_load()`)
or a workspace (`.jws_load()`).
}
